// src/frontend/src/app/api/models/daily-activity-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { ActivityStatus } from '../models/activity-status';
export interface DailyActivityDto {
  content: string;
  createdAt: string;
  id: string;
  isBlocked: boolean;
  isPublished: boolean;
  partnerId: string;
  referenceDate: string;
  status: ActivityStatus;
  theme: string;
  title: string;
}